[cli] migrator CLI handles init and generate commands
