+++ sea-orm-cli/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod cli;
-#[cfg(feature = "codegen")]
 pub mod commands;
 
 pub use cli::*;
-#[cfg(feature = "codegen")]
 pub use commands::*;
