+++ sea-orm-migration/src/cli.rs
@@ -1,12 +1,14 @@
 use clap::Parser;
 use dotenv::dotenv;
-use std::{fmt::Display, process::exit};
+use std::{error::Error, fmt::Display, process::exit};
 use tracing_subscriber::{prelude::*, EnvFilter};
 
 use sea_orm::{Database, DbConn};
-use sea_orm_cli::MigrateSubcommands;
+use sea_orm_cli::{run_migrate_generate, run_migrate_init, MigrateSubcommands};
 
 use super::MigratorTrait;
+
+const MIGRATION_DIR: &str = "./";
 
 pub async fn run_cli<M>(migrator: M)
 where
@@ -17,10 +19,17 @@ where
     let db = &Database::connect(&url).await.unwrap();
     let cli = Cli::parse();
 
-    run_migrate(migrator, db, cli.command, cli.verbose).await;
+    run_migrate(migrator, db, cli.command, cli.verbose)
+        .await
+        .unwrap_or_else(handle_error);
 }
 
-pub async fn run_migrate<M>(_: M, db: &DbConn, command: Option<MigrateSubcommands>, verbose: bool)
+pub async fn run_migrate<M>(
+    _: M,
+    db: &DbConn,
+    command: Option<MigrateSubcommands>,
+    verbose: bool,
+) -> Result<(), Box<dyn Error>>
 where
     M: MigratorTrait,
 {
@@ -49,15 +58,20 @@ where
     };
 
     match command {
-        Some(MigrateSubcommands::Fresh) => M::fresh(db).await,
-        Some(MigrateSubcommands::Refresh) => M::refresh(db).await,
-        Some(MigrateSubcommands::Reset) => M::reset(db).await,
-        Some(MigrateSubcommands::Status) => M::status(db).await,
-        Some(MigrateSubcommands::Up { num }) => M::up(db, Some(num)).await,
-        Some(MigrateSubcommands::Down { num }) => M::down(db, Some(num)).await,
-        _ => M::up(db, None).await,
-    }
-    .unwrap_or_else(handle_error);
+        Some(MigrateSubcommands::Fresh) => M::fresh(db).await?,
+        Some(MigrateSubcommands::Refresh) => M::refresh(db).await?,
+        Some(MigrateSubcommands::Reset) => M::reset(db).await?,
+        Some(MigrateSubcommands::Status) => M::status(db).await?,
+        Some(MigrateSubcommands::Up { num }) => M::up(db, Some(num)).await?,
+        Some(MigrateSubcommands::Down { num }) => M::down(db, Some(num)).await?,
+        Some(MigrateSubcommands::Init) => run_migrate_init(MIGRATION_DIR)?,
+        Some(MigrateSubcommands::Generate { migration_name }) => {
+            run_migrate_generate(MIGRATION_DIR, &migration_name)?
+        }
+        _ => M::up(db, None).await?,
+    };
+
+    Ok(())
 }
 
 #[derive(Parser)]
